refactor: Rename button component & fix lint errors
