+++ examples/jest-vtu-example/.eslintignore
@@ -1,1 +1,1 @@
-*.spec.js
+*.spec.js.snap
